+++ blockfrost-api/test/Cardano/Transactions.hs
@@ -161,7 +161,8 @@ transactionUtxosSample = [r|
       "tx_hash": "1a0570af966fb355a7160e4f82d5a80b8681b7955f5d44bec0dce628516157f0",
       "output_index": 0,
       "collateral": false,
-      "data_hash": "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
+      "data_hash": "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710",
+      "inline_datum": null
     }
   ],
   "outputs": [
@@ -178,7 +179,8 @@ transactionUtxosSample = [r|
         }
       ],
       "data_hash": "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710",
-      "output_index": 0
+      "output_index": 0,
+      "inline_datum": "19a6aa"
     }
   ]
 }
@@ -193,6 +195,7 @@ utxoInSample =
       , _utxoInputOutputIndex = 0
       , _utxoInputCollateral = False
       , _utxoInputDataHash = Just "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
+      , _utxoInputInlineDatum = Nothing
       }
 
 utxoOutSample :: UtxoOutput
@@ -202,6 +205,7 @@ utxoOutSample =
     , _utxoOutputAmount = sampleAmounts
     , _utxoOutputDataHash = Just "9e478573ab81ea7a8e31891ce0648b81229f408d596a3483e6f4f9b92d3cf710"
     , _utxoOutputOutputIndex = 0
+    , _utxoOutputInlineDatum = Just $ InlineDatum $ ScriptDatumCBOR "19a6aa"
     }
 
 transactionUtxosExpected =